+++ GlucoMan.Mobile/GlucoMan.Mobile/MealsPage.xaml.cs
@@ -30,41 +30,40 @@ namespace GlucoMan.Mobile
         private void FromClassToUi()
         {
             txtIdMeal.Text = bl.Meal.IdMeal.ToString();
-            txtChoOfMeal.Text = Safe.String(bl.Meal.CarbohydratesGrams.Text);
+            txtChoOfMeal.Text = Safe.String(bl.Meal.ChoGrams.Text);
             if (bl.Meal.TimeBegin.DateTime != Common.DateNull)
             {
                 dtpMealDateBegin.Date = (DateTime)Safe.DateTime(bl.Meal.TimeBegin.DateTime);
                 dtpMealTimeBegin.Time = ((DateTime)bl.Meal.TimeBegin.DateTime).TimeOfDay;  // - dtpMealDateBegin.Date;
             }
             txtAccuracyOfChoMeal.Text = bl.Meal.AccuracyOfChoEstimate.ToString();
-            cmbAccuracyMeal.SelectedItem = bl.Meal.QualitativeAccuracyOfChoEstimate;
 
             cmbTypeOfMeal.SelectedItem = bl.Meal.IdTypeOfMeal;
             cmbAccuracyMeal.SelectedItem = bl.Meal.QualitativeAccuracyOfChoEstimate;
             SetCorrectRadioButtons(); 
         }
-
         private void SetCorrectRadioButtons()
         {
+            // un-check all
             rdbIsBreakfast.IsChecked = false;
             rdbIsSnack.IsChecked = false;
             rdbIsLunch.IsChecked = false;
             rdbIsDinner.IsChecked = false;
 
+            // check the one that is enabled now (if any!) 
             if (bl.Meal.IdTypeOfMeal == TypeOfMeal.Breakfast)
                 rdbIsBreakfast.IsChecked = true;
             else if (bl.Meal.IdTypeOfMeal == TypeOfMeal.Snack)
                 rdbIsSnack.IsChecked = true;
-            else if (bl.Meal.IdTypeOfMeal != TypeOfMeal.Lunch)
+            else if (bl.Meal.IdTypeOfMeal == TypeOfMeal.Lunch)
                 rdbIsLunch.IsChecked = true;
             else if (bl.Meal.IdTypeOfMeal == TypeOfMeal.Dinner)
                 rdbIsDinner.IsChecked = true;
         }
-
         private void FromUiToClass()
         {
             bl.Meal.IdMeal = Safe.Int(txtIdMeal.Text);
-            bl.Meal.CarbohydratesGrams.Text = Safe.Double(txtChoOfMeal.Text).ToString();
+            bl.Meal.ChoGrams.Text = Safe.Double(txtChoOfMeal.Text).ToString();
             DateTime instant = new DateTime(dtpMealDateBegin.Date.Year, dtpMealDateBegin.Date.Month, dtpMealDateBegin.Date.Day,
                 dtpMealTimeBegin.Time.Hours, dtpMealTimeBegin.Time.Minutes, dtpMealTimeBegin.Time.Seconds);
             bl.Meal.TimeBegin.DateTime = instant;
@@ -76,11 +75,27 @@ namespace GlucoMan.Mobile
                 bl.Meal.IdTypeOfMeal = (TypeOfMeal)cmbTypeOfMeal.SelectedItem;
             else
                 bl.Meal.IdTypeOfMeal = TypeOfMeal.NotSet;
-            
+
             if (cmbAccuracyMeal.SelectedItem != null)
                 bl.Meal.QualitativeAccuracyOfChoEstimate = (QualitativeAccuracy)cmbAccuracyMeal.SelectedItem;
             else
                 bl.Meal.QualitativeAccuracyOfChoEstimate = QualitativeAccuracy.NotSet;
+
+            // since the combo has more options, it gets the priority 
+            // over the radiobuttons, but if the combo is in one of the 
+            // states represented by radiobutton, the radiobutton gets the priority
+            if ((TypeOfMeal)cmbTypeOfMeal.SelectedItem != TypeOfMeal.Other
+                && (TypeOfMeal)cmbTypeOfMeal.SelectedItem != TypeOfMeal.NotSet)
+            {
+                if (rdbIsBreakfast.IsChecked)
+                    bl.Meal.IdTypeOfMeal = TypeOfMeal.Breakfast;
+                else if (rdbIsSnack.IsChecked)
+                    bl.Meal.IdTypeOfMeal = TypeOfMeal.Snack;
+                else if (rdbIsLunch.IsChecked)
+                    bl.Meal.IdTypeOfMeal = TypeOfMeal.Lunch;
+                else if (rdbIsDinner.IsChecked)
+                    bl.Meal.IdTypeOfMeal = Common.TypeOfMeal.Dinner;
+            }
         }
         private void RefreshUi()
         {
@@ -113,9 +128,15 @@ namespace GlucoMan.Mobile
                 await DisplayAlert("Deletion not possible", "Choose the meal to delete", "Ok");
                 return;
             }
-            FromUiToClass();
-            bl.DeleteOneMeal(bl.Meal);
-            RefreshUi();
+            bool remove = await DisplayAlert(String.Format("Should we delete the meal begun at {0}, Id {1}?",
+                bl.Meal.TimeBegin.Text,
+                bl.Meal.IdMeal), 
+                "", "Yes", "No");
+            if (remove)
+            {
+                bl.DeleteOneMeal(bl.Meal);
+                RefreshUi();
+            }
         }
         private async Task btnSaveMeal_ClickAsync(object sender, EventArgs e)
         {
@@ -150,9 +171,8 @@ namespace GlucoMan.Mobile
         }
         private void btnDefault_Click(object sender, EventArgs e)
         {
-            bl.NewDefaults(); 
-            RefreshUi(); 
-        }
+            bl.NewDefaults();
+            FromClassToUi();         }
         private async void OnGridSelectionAsync(object sender, SelectedItemChangedEventArgs e)
         {
             if (e.SelectedItem == null)
